ultime fix per luogo ed evento
