+++ src/components/ItaliaTheme/View/Commons/Dates.jsx
@@ -1,5 +1,4 @@
 import { defineMessages, useIntl } from 'react-intl';
-
 import React from 'react';
 import moment from 'moment/min/moment-with-locales';
 import { rrulestr } from 'rrule';
@@ -56,7 +55,9 @@ const Dates = ({ content, show_image }) => {
       RRULE_LANGUAGE.dateFormatter,
     );
   }
-
+  console.log(content.start)
+  console.log(content.end)
+  console.log(moment.locale());
   return content ? (
     <>
       <div className="point-list-wrapper my-4 mb-5">
