+++ src/components/ItaliaTheme/View/Commons/Sponsors.jsx
@@ -61,32 +61,8 @@ const Sponsor = ({ item }) => {
  * @params {string} folder name where to find images.
  * @returns {string} Markup of the component.
  */
-const Sponsors = ({ content, folder_name, title }) => {
+const Sponsors = ({ sponsors, title }) => {
   const intl = useIntl();
-  const url = `${flattenToAppURL(content['@id'])}/${folder_name}`;
-  const searchResults = useSelector((state) => state.search.subrequests);
-  const dispatch = useDispatch();
-  useEffect(() => {
-    if (content?.items.some((e) => e.id === folder_name)) {
-      dispatch(
-        searchContent(
-          url,
-          {
-            'path.depth': 1,
-            sort_on: 'getObjPositionInParent',
-            metadata_fields: '_all',
-            fullobjects: true,
-          },
-          folder_name,
-        ),
-      );
-    }
-    return () => {
-      dispatch(resetSearchContent(folder_name));
-    };
-  }, [dispatch, content, url, folder_name]);
-
-  const sponsors = searchResults?.[folder_name]?.items || [];
   const sponsors_no_logos = sponsors.filter((sponsor) => !sponsor.image);
   const sponsors_logos = sponsors.filter((sponsor) => sponsor.image);
   return sponsors?.length > 0 ? (
