+++ theme/ItaliaTheme/Views/_common.scss
@@ -202,9 +202,9 @@ div.sticky-wrapper {
 
       .date-label {
         width: 1.4rem;
-        font-size: 1rem;
+        font-size: 0.7rem;
         text-align: center;
-        font-family: $font-family-monospace-light;
+        font-family: $font-family-monospace;
         font-weight: 500;
         text-transform: lowercase;
       }
